refactor: 패키지 수정, test: TC generator 생성 방법 수정, OrderControllerTest Api 문서화를 위한 테스트코드 추가 작성 (Location 추가 예정)
