+++ src/test/java/com/example/surveyapp/config/generator/OrderFixtureGenerator.java
@@ -1,10 +1,7 @@
 package com.example.surveyapp.config.generator;
 
-import com.example.surveyapp.domain.order.model.Order;
-import com.example.surveyapp.domain.order.model.OrderItem;
-import com.example.surveyapp.domain.product.domain.model.Product;
-import com.example.surveyapp.domain.user.domain.model.User;
-import org.springframework.test.util.ReflectionTestUtils;
+import com.example.surveyapp.domain.order.domain.model.Order;
+import com.example.surveyapp.domain.order.domain.model.vo.OrderItem;
 
 import java.util.List;
 
@@ -15,7 +12,7 @@ public class OrderFixtureGenerator {
     public static Order generateOrderFixture(Long userId) {
         OrderItem item = OrderItemFixtureGenerator.generatorOrderItemFixture();
 
-        Order order = Order.create(userId, List.of(item));
+        Order order = Order.of(userId, List.of(item));
 
         return order;
