+++ src/test/java/com/example/surveyapp/config/generator/OrderItemFixtureGenerator.java
@@ -1,18 +1,17 @@
 package com.example.surveyapp.config.generator;
 
-import com.example.surveyapp.domain.order.model.OrderItem;
-import com.example.surveyapp.domain.order.model.OrderItemPoints;
+import com.example.surveyapp.domain.order.domain.model.vo.OrderItem;
+import com.example.surveyapp.domain.order.domain.model.vo.OrderItemPoints;
 
 public class OrderItemFixtureGenerator {
 
     private static final Long PRODUCT_ID = 1L;
     private static final String PRODUCT_TITLE = "chicken";
-    private static final OrderItemPoints PRODUCT_PRICE = OrderItemPoints.create(2500L);
-
+    private static final OrderItemPoints PRODUCT_PRICE = OrderItemPoints.of(2500L);
 
     public static OrderItem generatorOrderItemFixture(){
 
-        OrderItem item = OrderItem.create(PRODUCT_ID,
+        OrderItem item = OrderItem.of(PRODUCT_ID,
                 PRODUCT_TITLE,
                 PRODUCT_PRICE);
         return item;
